change type of money to float

+++ Firebase/UserBoard.cs
@@ -4,14 +4,14 @@ using System.Collections.Generic;
 [Serializable]
 public class UserBoard
 {
-    public int balance;
+    public float balance;
     public int scene;
 
-    public UserBoard(int balance, int scene)
+    public UserBoard(float balance, int scene)
     {
         this.balance = balance;
         this.scene = scene;
-    }
+    }    
 
     public Dictionary<string, Object> ToDictionary()
     {
@@ -20,5 +20,5 @@ public class UserBoard
         result["scene"] = scene;
 
         return result;
-    }
+    }   
 }
